refactor: move custom button content helpers into shared utils
